==> Reviews/4_glassdoor_lda.ipynb <==
from nltk.tokenize import TweetTokenizer
from nltk.corpus import stopwords
import string
import contractions
from spacy.lang.en.stop_words import STOP_WORDS
from nltk.stem import WordNetLemmatizer 
from textblob import TextBlob
import re
import pandas as pd
import numpy as np
from tqdm.notebook import tqdm_notebook
tqdm_notebook.pandas()
import ipywidgets
import swifter
import gensim
from gensim import corpora
import pickle
import os

glassdoor_data_dir = "/ifs/projects/amirgo-identification/glassdoor_data/"
glassdoor_reviews = pd.read_csv(os.path.join(glassdoor_data_dir, "glassdoor_reviews_only.csv")
glassdoor_lda_model_dir = "/ifs/projects/amirgo-identification/glassdoor_lda_model/"

pros = [doc.split(' ') for doc in glassdoor_reviews['pros_cleaned'].dropna().tolist()]
dictionary = corpora.Dictionary(pros)
corpus = [dictionary.doc2bow(doc) for doc in pros]
pickle.dump(corpus, open(os.path.join(glassdoor_lda_model_dir, 'corpus.pkl'), 'wb'))
dictionary.save(os.path.join(glassdoor_lda_model_dir, 'dictionary.gensim')

%%time
NUM_TOPICS = 10

ldamodel = gensim.models.LdaMulticore(corpus, num_topics = NUM_TOPICS, id2word=dictionary, passes=5, workers=4)
ldamodel.save(os.path.join(glassdoor_lda_model_dir, 'model5.gensim'))

topics = ldamodel.print_topics(num_words=10)
for topic in topics:
    print(topic)

Topic Modeling Cons

cons = [doc.split(' ') for doc in glassdoor_reviews['cons_cleaned'].dropna().tolist()]
cons_dictionary = corpora.Dictionary(cons)
cons_corpus = [cons_dictionary.doc2bow(doc) for doc in cons]
pickle.dump(cons_corpus, open(os.path.join(glassdoor_lda_model_dir, 'cons_corpus.pkl'), 'wb'))
cons_dictionary.save(os.path.join(glassdoor_lda_model_dir, 'cons_dictionary.gensim'))

%%time
NUM_TOPICS = 10
ldamodel = gensim.models.LdaMulticore(cons_corpus, num_topics = NUM_TOPICS, id2word=cons_dictionary, passes=5, workers=8)
ldamodel.save(os.path.join(glassdoor_lda_model_dir, 'cons_glassdoor_model.gensim'))

topics = ldamodel.print_topics(num_words=10)
for topic in topics:
    print(topic)

Prediction on Survey Data

survey_data = pd.read_csv("/ifs/gsb/amirgo/spacespace/spacespace/Coco/analyses_data/survey_hr_topic_modeling.csv")
survey_data = survey_data.dropna(subset=['pros_cleaned', 'cons_cleaned'])
survey_data.reset_index(drop=True, inplace=True)

from gensim.models.ldamulticore import LdaMulticore
ldamodel = LdaMulticore.load(os.path.join(glassdoor_lda_model_dir, 'model5.gensim'))
pros = [doc.split(' ') for doc in survey_data['pros_cleaned'].tolist()]
doc_vectors = [ldamodel.id2word.doc2bow(doc) for doc in pros]
doc_topics = [ldamodel[doc_vector] for doc_vector in doc_vectors]
rows = []
for doc in doc_topics:
    curr_row = [0] * 10
    for topic, prop in doc:
        curr_row[topic] = prop
    rows.append(curr_row)
pros_df = pd.DataFrame(rows, columns=['pros_topic'+str(i) for i in range(1,11)])

ldamodel = LdaMulticore.load(os.path.join(glassdoor_lda_model_dir, 'cons_glassdoor_model.gensim'))
cons = [doc.split(' ') for doc in survey_data['cons_cleaned'].tolist()]
doc_vectors = [ldamodel.id2word.doc2bow(doc) for doc in cons]
doc_topics = [ldamodel[doc_vector] for doc_vector in doc_vectors]
rows = []
for doc in doc_topics:
    curr_row = [0] * 10
    for topic, prop in doc:
        curr_row[topic] = prop
    rows.append(curr_row)
cons_df = pd.DataFrame(rows, columns=['cons_topic'+str(i) for i in range(1,11)])


pros_df.reset_index(drop=True, inplace=True)
cons_df.reset_index(drop=True, inplace=True)
survey_data = pd.concat([survey_data, pros_df, cons_df], axis=1)


survey_data.to_csv('/ifs/gsb/amirgo/spacespace/spacespace/Coco/analyses_data/coco_glassdoor_topic_modeling.csv', index=False)

Glassdoor Document Proportions 

glassdoor_reviews = glassdoor_reviews.dropna(subset=['pros_cleaned', 'cons_cleaned'])

from gensim.models.ldamulticore import LdaMulticore
ldamodel = LdaMulticore.load(os.path.join(glassdoor_lda_model_dir, 'model5.gensim'))
# redoing this without loading saved corpus as the original models were not 
pros = [doc.split(' ') for doc in glassdoor_reviews['pros_cleaned'].tolist()]
doc_vectors = [ldamodel.id2word.doc2bow(doc) for doc in pros]
doc_topics = [ldamodel[doc_vector] for doc_vector in doc_vectors]
rows = []
for doc in doc_topics:
    curr_row = [0] * 10
    for topic, prop in doc:
        curr_row[topic] = prop
    rows.append(curr_row)
glassdoor_pros_df = pd.DataFrame(rows, columns=['pros_topic'+str(i) for i in range(1,11)])

ldamodel = LdaMulticore.load(os.path.join(glassdoor_lda_model_dir, 'cons_glassdoor_model.gensim'))
cons = [doc.split(' ') for doc in glassdoor_reviews['cons_cleaned'].tolist()]
doc_vectors = [ldamodel.id2word.doc2bow(doc) for doc in cons]
doc_topics = [ldamodel[doc_vector] for doc_vector in doc_vectors]
rows = []
for doc in doc_topics:
    curr_row = [0] * 10
    for topic, prop in doc:
        curr_row[topic] = prop
    rows.append(curr_row)
glassdoor_cons_df = pd.DataFrame(rows, columns=['cons_topic'+str(i) for i in range(1,11)])

glassdoor_pros_df.reset_index(drop=True, inplace=True)
glassdoor_cons_df.reset_index(drop=True, inplace=True)
glassdoor_reviews.reset_index(drop=True, inplace=True)
glassdoor_reviews.drop(['pros_cleaned', 'cons_cleaned', 'Unnamed: 0'], axis=1, inplace=True)

glassdoor_reviews = pd.concat([glassdoor_reviews, glassdoor_pros_df, glassdoor_cons_df], axis=1)
glassdoor_reviews.to_csv(os.path.join(glassdoor_data_dir, 'glassdoor_topic_prop.csv'), index=False)

glassdoor_pros_ldamodel = LdaMulticore.load(os.path.join(glassdoor_lda_model_dir, 'model5.gensim'))
pros_names = ['pros_topic'+str(i) for i in range(1,11)]
topics = glassdoor_pros_ldamodel.show_topics(formatted=False)
rows = []
for i in range(0, len(topics)):
    row = [pros_names[i]]
    word_prob_list = topics[i][1]
    for word, _ in word_prob_list:
        row.append(word)
    rows.append(row)

pros_top_words = pd.DataFrame(rows)
pros_top_words.set_index(0, inplace=True)
pros_top_words.to_csv(os.path.join(glassdoor_data_dir, 'glassdoor_pros_top_words_lda_k10.csv'))

glassdoor_cons_ldamodel = LdaMulticore.load(os.path.join(glassdoor_lda_model_dir, 'cons_glassdoor_model.gensim'))
cons_names = ['cons_topic'+str(i) for i in range(1,11)]
topics = glassdoor_cons_ldamodel.show_topics(formatted=False)
rows = []
for i in range(0, len(topics)):
    row = [cons_names[i]]
    word_prob_list = topics[i][1]
    for word, _ in word_prob_list:
        row.append(word)
    rows.append(row)
cons_top_words = pd.DataFrame(rows)
cons_top_words.set_index(0, inplace=True)
cons_top_words.to_csv(os.path.join(glassdoor_data_dir, 'glassdoor_cons_top_words_lda_k10.csv'))

glassdoor_reviews = pd.read_csv(os.path.join(glassdoor_data_dir, "glassdoor_reviews_only.csv"))
glassdoor_reviews = glassdoor_reviews.dropna(subset=['pros_cleaned', 'cons_cleaned'])
glassdoor_pros_df.reset_index(drop=True, inplace=True)
glassdoor_cons_df.reset_index(drop=True, inplace=True)
glassdoor_reviews.reset_index(drop=True, inplace=True)
glassdoor_reviews = pd.concat([glassdoor_reviews, glassdoor_pros_df, glassdoor_cons_df], axis=1)

glassdoor_original = pd.read_csv(os.path.join(glassdoor_data_dir, 'reviews_new_processed.csv'))

pd.set_option('max_colwidth', 400)
cons_t1 = glassdoor_reviews[['review_id','cons_cleaned', 'cons_topic1']].sort_values('cons_topic1', ascending=False).head(15)
cons_t2 = glassdoor_reviews[['review_id','cons_cleaned', 'cons_topic2']].sort_values('cons_topic2', ascending=False).head(15)
cons_t4 = glassdoor_reviews[['review_id','cons_cleaned', 'cons_topic4']].sort_values('cons_topic4', ascending=False).head(15)

cons_t1.merge(glassdoor_original[['review_id', 'cons']], on='review_id')[['cons_topic1', 'cons']]

cons_t2.merge(glassdoor_original[['review_id', 'cons']], on='review_id')[['cons_topic2', 'cons']]

cons_t4.merge(glassdoor_original[['review_id', 'cons']], on='review_id')[['cons_topic4', 'cons']]
